% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{availableModels}
\alias{availableModels}
\title{List all available chromWave models.}
\usage{
availableModels()
}
\value{
Function returns tibble of all chromwaveR included models.
}
\description{
\code{availableModels} returns all previously trained and available
chromWave nucleosome models.
}
\details{
\tabular{lll}{
 \strong{Model} \tab \strong{Organism} \tab \strong{Description}\cr
  invitro \tab sacCer \tab Model was trained on in vitro nucleosome data from Kaplan et al (2009) \cr
  invivo \tab sacCer \tab Model was trained on in vivo nucleosome data from Kaplan et al (2009) \cr
  TF-NUC \tab sacCer \tab Model was trained on MNase-seq data from Henikoff et al (2011)\cr
  hs_promoter \tab hs \tab Model was trained on promoter subset of MNase-seq data from Gaffney et al (2012) \cr
}
}
\examples{
availableModels()

}
