% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadChromWaveModel}
\alias{loadChromWaveModel}
\title{Load chromWave nucleosome model.}
\usage{
loadChromWaveModel(model.name)
}
\arguments{
\item{\code{model.name}}{Models include invitro, invivo, TF-NUC and hs_promoter.}
}
\value{
Function returns chromWave model as keras object.
}
\description{
\code{loadChromWaveModel} loads pre-trained nucleosome model from *.json/*.h5
files included in this package. The model input shape is altered by this function
to (None, None, 4) enabling chromosome wide predicitons.
Function serves as wrapper for python function load_chromwave_model.
Please check \code{availableModels()} for list of pretrained nucleosome models.
}
\examples{
## Load yeast in vivo nucleosome model
model <- loadChromWaveModel(model_name = 'invivo')

## Load human promoter nucleosome model
model <- loadChromWaveModel(model_name = 'hs_promoter')

}
