% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oneHotEncode}
\alias{oneHotEncode}
\title{Convert DNA sequence from string to one hot encoded matrix.}
\usage{
oneHotEncode(dna.strs)
}
\arguments{
\item{\code{dna.strs}}{DNAStringSet}
}
\value{
Function returns list of one-hot encoded DNAStrings.
}
\description{
\code{oneHotEncode} is a wrapper function for the
chromWave python function one_hot_encode.
Function converts DNA strings to one-hot encoded sequence matrices.
}
\examples{
# Convert DNAStringSet to list of one-hot encoded matrices
dna.strs <- rmdDNA()
oneHot.strs <- oneHotEncode(dna.strs)

}
