% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmdDNA}
\alias{rmdDNA}
\title{Function to create random DNA strings.}
\usage{
rmdDNA(n = 10, width = 500)
}
\arguments{
\item{\code{n}}{Number of rmd generated DNA strings.}

\item{\code{width}}{Length in bp of generated DNA strings.}
}
\value{
Function returns randomly generated DNAStringSet.
}
\description{
\code{rmdDNA} creates a DNAStringSet with \code{n} DNAStrings with
the width \code{width}.
}
\examples{
# Create 100 rmd DNA sequences of 1000bp lenght
rmdDNA(n = 100, width = 10^3)

}
